Refactor out some properties and helper methods of SizerBase to their own files in the partial class.
Move IsDragInverted to ContentSizer as it is specific to that class for now.
