+++ Microsoft.Toolkit.Uwp.UI.Controls.Layout/Sizers/ContentSizer/ContentSizer.Properties.cs
@@ -13,6 +13,21 @@ namespace Microsoft.Toolkit.Uwp.UI.Controls
     public partial class ContentSizer
     {
         /// <summary>
+        /// Gets or sets a value indicating whether the <see cref="ContentSizer"/> control is resizing in the opposite direction.
+        /// </summary>
+        public bool IsDragInverted
+        {
+            get { return (bool)GetValue(IsDragInvertedProperty); }
+            set { SetValue(IsDragInvertedProperty, value); }
+        }
+
+        /// <summary>
+        /// Identifies the <see cref="IsDragInverted"/> dependency property.
+        /// </summary>
+        public static readonly DependencyProperty IsDragInvertedProperty =
+            DependencyProperty.Register(nameof(IsDragInverted), typeof(bool), typeof(ContentSizer), new PropertyMetadata(false));
+
+        /// <summary>
         /// Gets or sets the control that the <see cref="ContentSizer"/> is resizing. Be default, this will be the visual ancestor of the <see cref="ContentSizer"/>.
         /// </summary>
         public FrameworkElement TargetControl
